Changes made to increase accuracy - Dragon score
Following the recent test on the LCS (16 Feb), the result was 80% accuracy. It increased the overall accuracy to 93.3% but nevertheless it had an error. I believe this change may increase accuracy. It also takes in the teams ability to score a dragon. It is added to the score and then decides a winner. 
I will test this for next week's LEC matches for diversity in testing. 
Test logs for this week will be available soon.