Test Logs - Week 2 of testing
Test logs for this week. Almost every prediction was correct. The only outlier was the match on 17 Feb (according to my time zone) 100 Thieves vs FlyQuest. I have to figure out the reason for its fault. For now, the script with the Dragon Update is the current version, which is working. I will try and update the script with aspects to increase accuracy.